Fixed bug where mutations due to insertions wouldn't highlight. Modified mutation support to be based off a slider from 0.05 to 0.5. Modified mutation support (same thing as mutation coverage) to be relative to the total amount of coverage at an arbitrary position, as opposed to a strict cut-off like it previously was

+++ iGenomics/BWT/BWT_MutationFilter/BWT_MutationFilter.h
@@ -47,7 +47,7 @@ extern int coverageArray[kMaxBytesForIndexer*kMaxMultipleToCountAt];//I--------G
     
     int fileStrLen;
 }
-@property (nonatomic) int kHeteroAllowance;
+@property (nonatomic) float kHeteroAllowance;
 - (void)setUpMutationFilterWithOriginalStr:(char*)originalSeq andMatcher:(BWT_Matcher*)myMatcher;
 - (void)resetFoundGenome;
 
@@ -56,6 +56,6 @@ extern int coverageArray[kMaxBytesForIndexer*kMaxMultipleToCountAt];//I--------G
 - (NSArray*)findMutationsWithOriginalSeq:(char*)seq;
 
 + (NSMutableArray*)filteredMutations:(NSArray*)arr
-                  forHeteroAllowance:(int)heteroAllowance insertionsArr:(NSArray*)insArr;
+                  forHeteroAllowance:(float)heteroAllowance insertionsArr:(NSArray*)insArr;
 + (NSMutableArray*)compareFoundMutationsArr:(NSArray *)arr toImptMutationsString:(NSString *)imptMutsStr andCumulativeLenArr:(NSMutableArray*)lenArr andSegmentNameArr:(NSMutableArray*)nameArr;
 @end
